+++ CHANGES.md
@@ -13,6 +13,9 @@ To be released.
  -  Added `BlockProtocolVersion` property to `IActionContext`.  [[#3228]]
  -  Changed `IAccountStateDelta.TransferAsset()` to require additional
     `IActionContext` parameter.  [[#3229]]
+ -  Changed `IAccountStateDelta.MintAsset()` and
+    `IAccountStateDelta.BurnAsset()` to each require an additional
+    `IActionContext` parameter.  [[#3231]]
 
 ### Backward-incompatible network protocol changes
 
@@ -33,6 +36,7 @@ To be released.
 
 [#3228]: https://github.com/planetarium/libplanet/pull/3228
 [#3229]: https://github.com/planetarium/libplanet/pull/3229
+[#3231]: https://github.com/planetarium/libplanet/pull/3231
 
 
 Version 2.1.0